back_end/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==4.0.0
django-environ==0.9.0
django-filter==23.2
djangorestframework==3.14.0
gunicorn==20.1.0
Markdown==3.4.3
psycopg2-binary==2.9.5
pytz==2023.3
sqlparse==0.4.3
